fs_init: mark whole page-chain of each file dirty.
On startup the dataflash is scanned for files and each referenced
page is marked dirty.  So far only the first page has been marked
as used however, i.e. if a new block had to be allocated the further
pages of each file haven't been protected and could be re-used and
thus overwritten.