httpd: reorder sent headers so we can omit sending Content-Length.
We don't send a Content-Length header if the size of the file,
as reported by the VFS, is zero.  This is handy for the (yet to
be written) camera module VFS implementations that don't know
the JPEG's size in advance.

send_length_if [VFS_TEENSY]: use %d instead of %ld
state->len is uint16_t in that case (uint32_t otherwise).