Add Buderus EMS heating bus implementation.
This implementation is based on knowledge gathered from this forum
thread: http://www.mikrocontroller.net/topic/141831 - it should reflect
everything mentioned there.

The expected hardware is basically this schematic plus a few LEDs:
http://www.mikrocontroller.net/topic/141831#1976381

The implementation was based on the YPort protocol implementation. It
uses a (configurable) hardware UART and Timer 2.