Calculate the max length ECMD commands properly
Each ECMD command that arrives via http is url encoded.

If a command contains a lot of whitespaces/other escaped chars, it will
not fit into maxlen although the unescaped command is smaller than
ECMD_INPUTBUF_LENGTH.
The maximum length of a string that will be url encoded is three times the
length of the string if every character needs to be encoded, like in a
string that contains only whitespaces.

closes #427

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>