MQTT module
The static configuration of the connection data of
the MQTT broker is from now on the standard case,
comparable to other protocols.
The registration of any module with the MQTT module
is now done with the well-known Ethersex meta mechanism.
The generated configurations are stored in the flash
and save RAM. Registration at runtime is still possible,
but not the preferred way.